+++ libcircuit_reliability/src/simulations_handling/spectre_handlers/variability_spectre_handler.hpp
@@ -109,6 +109,7 @@ public:
 	bool get_plot_scatters(){ return plot_scatters; }
 	bool get_plot_transients(){ return plot_transients; }
 	void set_plot_transients( bool plot_transients) { this->plot_transients = plot_transients; }
+	bool get_export_metric_errors(){ return export_metric_errors; }
 	void set_export_metric_errors( bool export_metric_errors) { this->export_metric_errors = export_metric_errors; }
 	bool get_plot_last_transients(){ return plot_last_transients; }
 	void set_plot_last_transients( bool plot_last_transients) { this->plot_last_transients = plot_last_transients; }