+++ libcircuit_reliability/src/simulations_handling/spectre_simulations/montecarlo_critical_parameter_value_simulation.cpp
@@ -74,6 +74,9 @@ void MontecarloCriticalParameterValueSimulation::RunSimulation( ){
 	unsigned int threadsCount = 0;
 	critical_parameter_value_simulations_vector.set_group_name("critical_parameter_value_simulations_vector");
 	critical_parameter_value_simulations_vector.ReserveSimulationsInMemory( montecarlo_iterations );
+	#ifdef SPECTRE_SIMULATIONS_VERBOSE
+		log_io->ReportThread( "Total montecarlo to be simulated: " + number2String(montecarlo_iterations) + ". Max number of montecarlo threads: " + number2String(max_parallel_montecarlo_instances), 1 );
+	#endif
 	while( threadsCount<montecarlo_iterations ){
 		// wait for resources
 		WaitForResources( runningThreads, max_parallel_montecarlo_instances, mainTG, threadsCount );